feat: namespace LoadsAndAuthorizesResource to BulletTrain::LoadsAndAuthorizesResource
