+++ bullet_train-super_load_and_authorize_resource/.gitignore
@@ -1,5 +1,6 @@
 /.bundle/
 /doc/
+/Gemfile.lock
 /log/*.log
 /pkg/
 /tmp/
