+++ bullet_train-super_load_and_authorize_resource/bullet_train-super_load_and_authorize_resource.gemspec
@@ -23,4 +23,6 @@ Gem::Specification.new do |spec|
   end
 
   spec.add_dependency "rails", ">= 6.0.0"
+
+  spec.add_development_dependency "pry"
 end
